Hold money in a Long, because so'm do not fit in an Int
The cart stopped loading: "Failed to parse int for input '2688000000' at
path: $.items[0].line_total". Sixteen of a 168 000 000 so'm watch is
2 688 000 000, and a Kotlin Int stops at 2 147 483 647 — so the line
total was not a wrong number, it was a number the response could not be
read past at all, and the whole screen fell back to an error state.

Capping quantity at the shelf, which landed earlier today, does not
reach this. Twenty-five of that watch is 4.2 billion. The problem is the
type, not the count: an Int is too small for a currency whose everyday
amounts run to nine figures.

Twenty-four money fields widened — price, old price, unit price, line
total, subtotal, discount, delivery fee, total — with the formatters
taking Long and keeping an Int overload for the call sites that hold a
count. Percentages and counts stay Int.

The backend was never affected: Python integers do not overflow. Nor was
iOS, where Int is 64-bit — checked on the machine rather than assumed.

One thing to watch when reading this diff: `total` means two different
things in this codebase. A cart total and an order total are money; a
page total, a review count and a photograph count are not. The first
pass widened all of them, and the compiler caught it — plural resources
want an Int, which is how a review count gave itself away.

+++ android/app/src/main/java/uz/minibozor/core/design/component/Product.kt
@@ -279,8 +279,8 @@ fun MbDiscountPill(
  */
 @Composable
 fun MbPriceRow(
-    price: Int,
-    oldPrice: Int? = null,
+    price: Long,
+    oldPrice: Long? = null,
     discountPercent: Int? = null,
     modifier: Modifier = Modifier,
     priceStyle: TextStyle = MbTheme.type.price,
@@ -385,8 +385,8 @@ fun MbStars(rating: Double, modifier: Modifier = Modifier, size: Dp = 15.dp) {
  */
 @Composable
 fun MbHeroPrice(
-    price: Int,
-    oldPrice: Int? = null,
+    price: Long,
+    oldPrice: Long? = null,
     discountPercent: Int? = null,
     /**
      * A step smaller, for the bar at the top of the product page.
@@ -541,8 +541,8 @@ fun MbCartButton(
 @Composable
 fun MbProductTile(
     title: String,
-    price: Int,
-    oldPrice: Int?,
+    price: Long,
+    oldPrice: Long?,
     discountPercent: Int?,
     imageUrl: String?,
     isFavorite: Boolean,
@@ -573,8 +573,8 @@ fun MbProductTile(
 @Composable
 fun MbDealTile(
     title: String,
-    price: Int,
-    oldPrice: Int?,
+    price: Long,
+    oldPrice: Long?,
     discountPercent: Int?,
     imageUrl: String?,
     onClick: () -> Unit,
@@ -616,8 +616,8 @@ fun MbDealTile(
 @Composable
 private fun ProductTileBody(
     title: String,
-    price: Int,
-    oldPrice: Int?,
+    price: Long,
+    oldPrice: Long?,
     discountPercent: Int?,
     imageUrl: String?,
     isFavorite: Boolean,
@@ -716,8 +716,8 @@ private fun ProductTileBody(
 @Composable
 fun MbRailTile(
     title: String,
-    price: Int,
-    oldPrice: Int?,
+    price: Long,
+    oldPrice: Long?,
     discountPercent: Int?,
     imageUrl: String?,
     onClick: () -> Unit,
@@ -843,7 +843,7 @@ fun MbLineItem(
     title: String,
     imageUrl: String?,
     meta: String,
-    price: Int,
+    price: Long,
     quantity: Int? = null,
     modifier: Modifier = Modifier,
     onClick: (() -> Unit)? = null,
